fluent-bit: Add new default metric for eMMC health information
Update the fluent-bit.conf to send new data related to eMMC health.
This metric will send 2 pieces of information provided by a simple script.

The first piece of data is related to eMMC MLC lifetime estimation.
The second is related to eMMC reserved blocks EoL status.
This information will be sent in this format:

{"emmc_life_time_est_typ_a":"10","emmc_pre_eol_info":"1"}

Related-to: TOR-3273

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>