network-configuration: drop recipe and its references
Currently network-configuration only being used by qemuarm64 machine,
for the reason that we run QEMU with TAP network device and that
requires a static IP.

But we dont have to maintain a extra recipe for setting a static IP
which is managed by systemd-networkd. networkmanager can do that as
well.

Introduce a network.nmconnection.in for qemuarm64 to set up a static
IP in networkmanager.

Related-to: TOR-3336

Signed-off-by: Ming Liu <ming.liu@toradex.com>