torizon.inc: remove BBMASK from meta-freescale/cairo append
Having this BBMASK makes it impossible to have DISTRO_FEATURES
such as x11 or wayland on iMX8, which is an issue for users of
torizon-minimal.

As the append is imxgpu-specific, it shouldn't affect any other
machines. It also shouldn't affect the other torizon builds as
no package from those depends/requires cairo.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>