plymouth: drop pango and some themes
plymouth has a dependent chain like:
```
plymouth -> pango -> cairo -> virtual/egl, virtual/egl2
```

When PACKAGECONFIG 'pango' is enabled, it leads to all the above
packages installed to initramfs-ostree-torizon-image, as a result, the
initramfs size is 32M for apalis-imx8 machines.

While PACKAGECONFIG 'pango' is needed only for text rendering of user
prompts (allowing user to input passwd and so on), we dont really need
it, let's drop it.

Also we are only using spinner theme, let's drop the other themes.

By doing this, the initramfs size for a apalis-imx8 machine is reduced
to 16M from 32M.

Fix: https://github.com/torizon/meta-toradex-torizon/issues/16

Signed-off-by: Ming Liu <ming.liu@toradex.com>
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>