initramfs-framework: add modules to support plymouth on AM62
For AM62, plymouth wasn't able to show the splash screen due to the
framebuffer not being ready on initramfs.

So we needed to add tidss module (and it's dependencies), pwm and DSI
bridges modules to make it work with our DSI to LVDS adapter. With this
modification, it also made required that we insert the necessary modules
for the DSI to HDMI adapter to work, thats why we also see display_connector
and the lontium module.

fixes #30

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>