torizon.bbclass: amendments to prevent error messages during boot
Pass extra parameters in the kernel command line when composefs is
enabled in order to disable the "systemd-gpt-auto-generator"; this
generator is responsible for dynamically producing systemd units for GPT
partitions it finds on disk/storage. Since Torizon OS uses MBR
partitions, this shouldn't cause any side-effects and it avoids error
messages due to the generator not being able to determine the underlying
storage device when composefs is in use.

Resolves #32

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>