torizon-signed.bbclass: add override "cfs-support"
At the moment, this basically replaces override "torizon-signed" by one
named "cfs-support". The idea is to replace the former by two overrides;
the one being introduced now simply enables the use of composefs and the
one to be added later would be responsible for enforcing integrity and
authenticity on top of composefs.

Resolves #32

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>