ostree: switch to new composefs configuration method
Instead of enabling and configuring composefs directly on the ostree
repository, do it via the prepare-root.conf file (which is part of a
commit tree); the previous method will likely be dropped. The new method
available in ostree 2024.4 has the advantage of allowing the
configuration itself to be modified upon an update which wasn't possible
with the previous method.

Here we also enable the transient /etc on ostree in case composefs
support is enabled (secure boot case).

Resolves #37

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>