composefs-fsverity: add ed25519 key pair
Add key pair to support composefs fsverity.

The key pair is generated according to:
https://github.com/ostreedev/ostree/blob/main/tests/libtest.sh#L760

The private key should be stored in build system and will be refered
when "ostree --sign" runs.

The public key should be stored in initramfs and will be refered when
when "ostree deploy" runs at runtime.

Reference about how to deal with ed25519 keys:
https://github.com/ostreedev/ostree/blob/main/docs/composefs.md

Related-to: TOR-3378

Signed-off-by: Ming Liu <ming.liu@toradex.com>