initramfs-framework: automatically enable fsverity on first boot
Add logic to the ramdisk in order to enable fsverity when required by
the system configuration (i.e. when booting a secure boot image with
rootfs signing enabled) and detected that fsverity is still not enabled.
In practice this means the process would be done upon the first boot of
a device after installing the OS with Toradex Easy Installer since the
installer image is not capable of keeping fsverity information (as being
a simple tarball of the sysroot).

The approach being followed here has the advantage that it does not
require the build machine, the installer and the installer image to have
any knowledge of fsverity. The disadvantage is the extra boot time on
the first boot after initial installation; on a verdin-imx8mm the extra
time is around 2.5 minutes.

Related-to: TOR-3379

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>