recipes-image: introduce ti-img-rogue-driver as dependency
Commit c90cb38 dropped Pango as a dependency, which dropped the
virtual/egl dependency which made the out-of-tree GPU stack for
AM62 to never actually make into the final image.

This was detected using bisect and discovered due to the missing
firmware files in /usr/lib/firmware in the rootfs, more specifically
the rgx* binary blobs that the kernel module loads into the PVR core,
which is done at runtime when the kernel module gets exercised with
the correct IOCTLs (done by the libdrm component of the mesa-pvr
userspace stack, present in the Torizon OS reference containers).

This commit re-introduce this dependency, much like b40b911.

Closes #49.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>