docker: transition from docker-ce to docker-moby
docker-ce recipe has been dropped from meta-virtualization by commit
7ed351db, so we now use docker-moby instead.

Drop a patch where a path was added for a default config file
for the cli. This patch needed to be reworked, and since this seems to
be something we inherited from meta-lmp and we're not using, I'll just
remove this patch.

And collpase the torizon-docker.inc into the .bbappend recipe.

Closes #78

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>