initramfs-framework: remove unnecessary kernel modules from ramfs
After the bump to Kernel 6.6 we got a build error that some kernel
modules could not be found. After checking TI Kernel config, either
those configs were now builtin or non-existent, so they were removed.

Plymouth (boot splash screen) was tested and is working for both DSI to
LVDS and DSI to HDMI.

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>