torizon.inc: set expected bootargs when tdx-signed is employed
Here we are moving the expected bootargs from meta-toradex-security to
the present layer; the bootargs in the former layer are appropriate for
booting Toradex BSP images.

Setting the bootargs is relevant when the kernel command-line protection
feature provided by class tdx-signed (layer meta-toradex-security) is
enabled. Here the default settings from the class are overriden to make
them appropriate for booting Torizon OS images.

Related-to: TOR-3558

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>