recipes-sota: rac: update recipe
This release of RAC includes the possibility of receive
messages through D-Bus to poll the server for new sessions.
These messages are sent by the tzn-mqtt bridge (now included
as a dependency for the remote-access service) and thus we
can significantly increase the intervals to check the HTTP
server, reducing bandwidth.

Still, we keep the periodically checking the HTTP server
regardless of new messages on D-Bus as a fallback solution.

This default configuration is an example and can be changed
by users to fit their particular needs, specially in the case
where an applicaiton is bandwidth sensitive.

Note that this change also moves the rac repository to the
Torizon organization and also slightly updates the comment
in regards to AUTOREV.

Related-to: TOR-3522

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>