bl_actions.sh: add retry loops to critical operations
Add retry logic to the function responsible for determining the current
active boot partition. This deals with the fact that sometimes the
operation fails; the failure seems to come from the kernel.

Additionally, add retry logic to the function responsible for switching
the active boot partition. In this case, it was found that the operation
does not throw any errors even though it did not perform the expected
operation. To deal with this, we read back the current partition and
retry the switching if the current partition does not match the expected
one. Again the problem seems to come from the kernel.

These changes shouldn't be necessary and the proper solution would be at
the BSP level but keeping them should only add to robustness overall,
i.e. we don't necessarily need to revert the present commit once fixes
are in place on the BSP.

Related-to: TOR-3613

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>