composefs: backport composefs from meta-openembedded
Now that composefs is part of meta-openembedded, we can drop the recipe
we maintain in favor of the upstream one.
Since this is only available on master branch, I've backported this
recipe as is to our layer, renamed from 'composefs-tools' to 'composefs'
to be equivalent.

Related-to: TOR-3476
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit ff82d54b85af4eee2df7f6a974b78094a8cbc53c)

The cherry-pick has been changed to remove the backport of the recipe,
since on master we will get this from oe-core upstream instead.

Signed-off-by: Hiago De Franco <hiago.franco@toradex.com>