fluentbit: add Torizon specific customization on a separate bbappend
In this append we install Torizon specific configurations and we
correctly configure the build parameters.

Due to the fact that meta-openembedded is using PACKAGECONFIG, some
unused configs are being set by default to Off. This is causing some
build issue on fluentbit, and I believe it's happening because some of
the configs PACKAGECONFIG are explicitly turning Off might be On by
default.
So in order to fix this issue, and to make this fluent-bit build closer
to what we had, I've copied to this .bbappend the EXTRA_OECMAKE configs
we had, and removed the extra ones added by PACKAGECONFIG. This way we
end up with the same flags in EXTRA_OECMAKE we had before.

Related-to: 3476

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit a3625de294cb4bbd7e118dafd407b62b21df8d71)
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>