u-boot: add patch to save kernel_image_type from previous deployment
This is part of solving a bug which stops the rollback feature from
working when an update changes the kernel image type. It's worth
mentioning that the approach here is similar to the one done with the
fdtfile in which the information from the previous deployment is saved
in uEnv.txt and it similarly only works with two deployments, i.e. it
does not generalize to three or more.

Related-to: TOR-3563

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>