u-boot: Add mechanism to programmatically set fuses for HAB/AHAB (i.MX only)
This mechanism allows the programming of HAB/AHAB fuses via u-boot variables.
The following u-boot variables can be set for this mechanism to work:

* fuse_prog_list: An ordered, space separated list of hex values to be programmed into
the fuses.

* fuse_prog_close: A boolean 1 or 0 for whether the device should be closed for HAB/AHAB.
This variable is optional and if it does not exist it's the same as if it was 0.

* fuse_dry_run: Boolean 1 or 0. Causes the logic to do a dry run of the actions without
actually programming any fuses.

As feedback this mechanism will set the following environment variables.

* fuse_val_N: This contains the hex value currently programmed in the Nth
HAB/AHAB fuse on this device.

* fuse_status: This can have 3 states, "pending" means there is ongoing fuse programming
to be done, "failed" means there was an error during fuse programming, and "success" means
fuse programming was completed without any errors.

Related-to: TOR-3634

Signed-off-by: Jeremias Cordoba <js.cordoba8321@gmail.com>