initramfs-framework: sync rootfs script with open-embedded
Synchronize rootfs script with the one from open-embedded. With this the
only difference between them is that the version in our layer has a call
to tool `e2fsck` if the tool is available. Worth noting though is that
the initial ramdisk of Torizon OS does not include that tool which means
the call is mostly a no-op at the time of this writing.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>