u-boot-distro-boot: fix do_compile task to generate fuse bank words
The MR that added the fuse bank words [1], also added an issue on
do_compile task [2]. Since the arithmetic operator $((...)) is not
supported by bitbake inside a $(...) command expression. So 'let' was
used to update the for loop index, which is not supported by POSIX
shell.

So the fix was to add another task, 'get_fuse_bank_word', made in python
and scheduled to run before 'do_compile'. This way, this task processes
the BANK_WORD variable, produces the u-boot script string to be replaced
later and store it inside 'FUSE_BANK_WORDS'. Then we go into our
'do_compile:append' and simply replace our u-boot script with the newly
generated 'FUSE_BANK_WORDS.

Also all bank word variables are now a python tuple, so that we can
process them.

[1] https://github.com/torizon/meta-toradex-torizon/pull/190
[2] https://github.com/torizon/meta-toradex-torizon/pull/190/files#diff-4f692e15b504602e32946913e600028ae1ec2e65cfcccfd072283ffd2539c1edR135

Related-to: TOR-3663

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>