u-boot-distro-boot: fix 'get_bank_word_internal' with empty strings
For unsupported machines, BANK_WORD is an empty string and since this
was not being handled on this function we had an exception in this
function that caused the build to fail.

Now if BANK_WORD is empty, we return early with "" as well.

Related-to: TOR-3664

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit dca759e6c5ce6aa0e97cefd7cfd03224ea079bde)
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>