layer.conf: change meta-toradex-torizon layer dependency
Since it is now possible to use this layer to build Common Torizon with
other BSP's, it makes no sense to have this layer dependent on Toradex
and freescale layers.
So the dependency was changed to reflect the layer that really need to
be present so that 'meta-toradex-torizon' can be built.

Related-to: TOR-3661

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>