base-passwd: change base-passwd bbappend to match any version
Our bbappend recipe was appending only to a specific version of
base-passwd recipe, and since this recipe was bumped on OE upstream, we
were missing our group patched and having a build failure.

Backported from Scarthgap the 2 patches that add the necessary groups
and also corrects some GIDs to the expected values on Torizon.

Also removed unused patches input.patch and kvm.patch.

Related-to: TOR-3675

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>