linux-torizon.inc: make preempt-rt commit dates deterministic
This is an attempt to solve an issue with the PREEMPT-RT builds where
the build of the kernel (along with device-trees and overlays) does not
match the build of the kernel modules. Specifically, what does not match
is the string responsible for defining the kernel ABI version which in
turn also defines the name of the directory containing the kernel and
kernel modules.

To try and solve the issue, here we modify the do_patch task to make it
forcefully set the commit date of the last commit which is supposed to
be the commit applying the RT patches. Normally applying those patches
produces different hashes every time because the said patches do not
have date information in then leading the kgit-s2q tool to use the
current date on the commit produced by the patching operation.

Related-to: TOR-3656

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>