imx95: install mali firmware in the rootfs
We need the firmware as the kernel module expects it in specific places of the
root filesystem, so it cannot be dynamically deployed from containers as the
rest of the graphical stack is.

On the imx8 and am62 for example the firmware is installed via 'driver' recipes
(kernel-module-imx-gpu-viv and ti-img-rogue-installer), on the imx95, however,
it comes from userspace portion, thus we need the bbappend to only install the
firmware, nothing else (no libEGL, libGLES etc...).

This follows the same logic as commits a7ec4b08 and b40b9114.

Related-to: TCCP-879

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>