python/python3-urllib3: drop bbappend
The bbappend was applying a patch and removing dependencies only when
building the package for a target (i.e. non-native build). However, with
scarthgap, python3-urllib3 is no longer installed on the target in our
images (torizon-{minimal,docker,podman)) which means the append didn't
do any good for Toradex images and could potentially cause issues for
other users of the layer which in fact seems to be the case (PR
https://github.com/torizon/meta-toradex-torizon/pull/160).

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>