am62xx-evm: add OS rollback support
Add support for OS rollback by enabling the U-Boot environment to be
saved in a persistent storage. In this case the environment will be
saved in the first boot hardware partition of the eMMC
(/dev/mmcblk0boot0).

The values of CONFIG_ENV_OFFSET and CONFIG_ENV_SIZE were taken from this
tentative patch to U-Boot:

https://lore.kernel.org/all/5cda75ec-f7d4-bfe4-b23c-87394c19175b@ti.com/T/

In uEnv.txt.in, the uenv_extra_configs function now runs after
check_rollback_needed. This is done so that any auxiliary
functions/configs can also be applied during a rollback boot. In fact,
this is needed for this board, as uenv_extra_configs executes a function
to correct the fdtfile variable and it should always be executed,
whether the board is in rollback state or not.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>