initramfs-framework: add module to support Plymouth on AM62x/AM62P EVMs
For AM62x and AM62P Starter Kit EVMs the splash screen wasn't
appearing during boot on the HDMI output.

Both EVMs use a SiI9022A bridge chip to generate the HDMI signal, so the
corresponding driver should be loaded at the same stage where Plymouth
is initialized, which is in initramfs in our case.

The driver is built and loaded as a kernel module named sii902x.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>