conf: distro: xenomai.inc: unify shared distro data
Some distro data was in the intel-corei7-64.inc file, which is specific
to the machine. This makes it confusing to adjust config that is
applicable to the Xenomai distros.

Some of the configs will still be applicable only to the machine-distro
couple, but the reason for the overrides is the distro, so it sounds
more suitable to leave the configs in a distro file.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>