uEnv.txt: replace usage of tdx_is_closed by tdx_secboot_get
Replace the usage of the "tdx_is_closed" command by the more generic
"tdx_secboot_get", also created by Toradex. The new command allows the
reading of secure boot related properties directly into an environment
variable simplifying the logic.

The idea behind switching to this new command is to concentrate all
secboot-related decisions on information provided by it.  This
simplifies maintenance because we also concentrate the information
gathering on a single place in the U-Boot code. Plus we don't need to
review the commad whitelist whenever new information is needed by the
boot script to take its decisions.

Related-to: TOR-3657

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>