bootcommand.cfg: fix possible issue while sourcing boot.scr
U-Boot has some default addresses and variable it uses with some
commands.
In this scenario, we're first loading our boot.scr and then sourcing it,
like so:
=> ext4load ${devtype} ${devnum}:1 ${loadaddr} /boot.scr && source

We're explicitly loading the script into 'loadaddr' address, and on
source we're assuming the same variable will be utilized.
This assumption is wrong in U-Boot!
For the load command, if nothing is passed, it will use 'loadaddr' if it
exists and if not it will use the address of CONFIG_SYS_LOAD_ADDR.
While source, if nothing is passed, will use CONFIG_SYS_LOAD_ADDR
directly. And in this case, the values of 'loadaddr' and
CONFIG_SYS_LOAD_ADDR are different, which caused issues while trying to
source the script.

To avoid these issues in the future, we now explicit the address we're
using.

Related-to: TOR-3729

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>