fix: am62xx.inc is used by meta-ti and we were overriding it
On the Common Torizon AM62x machine refactor, the common .inc file was
labeled 'am62xx.inc', which happens to be the same name of an .inc file
used by 'meta-ti', and due to our layer precedence, this include was
being used instead of meta-ti's, resulting in a build error.

So the file is now renamed to 'common-am62xx.inc', and was already
changed on the machine .incs that use it.

Related-to: TOR-3774

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>