am62xx-evm.inc: remove KERNEL_DEVICETREE hard assignment
KERNEL_DEVICETREE is already being set in the MACHINE conf file on
meta-ti-bsp, so we don't need to set it in our layer.

However, we do need to hard assign OSTREE_DEVICETREE, as otherwise the
ostree-kernel-initramfs recipe fails when building 'torizon-docker':

cp: cannot stat
'.../deploy/images/am62xx-evm/k3-am625-phyboard-lyra-1-4-ghz-opp.dtbo':
No such file or directory

When setting KERNEL_DEVICETREE the TI BSP includes an overlay that
isn't present in neither the ti-linux-6.6.y nor on the ti-linux-6.12.y
branches from the TI downstream kernel source. This doesn't appear to
cause any problems during the kernel and overlay build steps, apparently
the non-existent DTs/DTOs are ignored and not built.

By default OSTREE_DEVICETREE will have the same value as
KERNEL_DEVICETREE if using sota.bbclass, which is the case without
the hard assignment.

The ostree-kernel-initramfs recipe uses OSTREE_DEVICETREE to copy the
DTBOs to their correct path in the final image, and it will fail if any
file referenced in the variable doesn't exist.

To avoid this issue, we can set OSTREE_DEVICETREE manually in our layer
to only have overlays that are present in the kernel sources.

Related-to: TOR-3782

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>