am62xx-evm: unset overlays being applied during boot
Remove the hard assignment to OSTREE_DEVICETREE_OVERLAY_BOOT in the
MACHINE inc file, which results in no overlays being applied during the
kernel boot process.

The image boots fine without the overlays, so we don't need to have them
applied by default. To be clear, the .dtbo files are still present in
the final image, they're just not applied by default during boot.

The OSTREE_DEVICETREE_OVERLAY_BOOT assignment has been commented with
instructions on how to apply the overlays if the user wishes to do so.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>