initramfs-framework: add module ite_it66121 only for BeagleY-AI
With the added support for BeagleY-AI, module 'ite_it66121' was added to
initramfs to support splashscreen. The thing is that this is being added
to all TI devices, even the ones using Toradex's BSP, which does not
have this module available.

So, in order to fix our build issues, we only add 'ite_it66121' for
BeagleY-AI machines.

Related-to: TOR-3781

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>