classes: fix torizon-signed generation
Prevent generation of stamp file for the "do_image_ostreecommit"
task. This is expected to solve an issue with the generation of the
torizon-signed images in which the key stored into the ramdisk may
not match the key actually used for signing the composefs digest.

Related-to: TOR-3829

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>
(cherry picked from commit 9d1905cbe40bea2dd63baf1b0c08a8707aa9d76b)