u-boot-distro-boot: introduce BOOTARG_ROOT override
In Torizon U-Boot passes the partition name to Linux via bootargs to
locate the root partition by its label. In STM32MP2-DK and
STM32MP2-EV1 can be 2 boot media: the on-board eMMC and the removable
SD-card, and if the system is installed on both of them, then
determining the root partition by name may not work correctly, since
there will be 2 partitions with the same name.

This commit introduces a new option in u-boot-distro-boot to override
the default root argument to allow an alternative method of
determining the root partition, such as by UUID.

Signed-off-by: Vladimir Skvortsov <vskvortsov@emcraft.com>