initramfs-framework: add irq_imx_irqsteer module to iMX8 initramfs
Due to BSP's change on linux-toradex-kconfig [1], a regression was
introduced on Torizon OS where the splash screen was no longer appearing
during boot on iMX8 devices (Colibri and Verdin).

This was due to the config CONFIG_IMX_IRQSTEER now being set to module,
so it would be missing during splash initialization, on initramfs.

Now we add the module to our initramfs for it to be inserted before
initializing plymouth.

[1] https://git.toradex.com/cgit/linux-toradex-kconfig.git/commit/?h=main&id=2bbba4e35cc4ac9f5ee0a6a44b5a6aa2a60324ba

Related-to: TOR-3803, TOR-3807, TOR-3837

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>