image_type_torizon: pack up files needed to re-sign OS images with TCB
As part of the effort to add in TorizonCore Builder the feature to
re-sign existing signed Torizon OS images, TCB needs to recreate the
bootloader container (imx-boot/flash.bin) from its constituent parts.

These files, composed of various U-Boot related binaries (SPL, U-Boot
proper, U-Boot DTB, ATF, firmwares, etc.), need to be easily available
in the image to be used by TCB. This TCB feature re-signs the kernel
fitImage as well, so it would also need to be made available.

This can be achieved by packing all required files in a tarball and
including it when building the Easy Installer image. By default this
will only be done when building images for supported machines that have
TDX_IMX_HAB_ENABLE set to 1. This variable is automatically set when
inheriting the 'tdx-signed' class from meta-toradex-security.

Users can also directly use TCB_SIGNING_SUPPORT to enable or disable the
above behavior.

Currently the only supported machine is the Verdin iMX8M Plus. Other
machines are planned to be supported in the future.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>