setup-environment-toradex: change CycloneDX SBOM export directory
In order for the SBOM artifacts to be published to artifactory, they
need to be present in the 'deploy/images/<MACHINE>' directory.
So, we change the SBOM names and path for them to be deployed in the
directory we want.

As soon as we have a more permanent solution in place, we can revert
this.

Related-to: TOR-3899

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit 8ea2d9f8ff9af75954cef88afab8484b7e5e8fae)