common: standardize hostnames
Hostnames should closely reflect the machine names. One of the main
reasons is clarity not only within Yocto but also due to the device
fields that are filled up in Torizon Cloud using the hostname as base.

This MR sets a precedent for new common machines hostnames, with the
following format: <machine-name>-<optional-qualifier>.

Note that <machine-name> itself might be a specific format with short
dashes and optional qualifiers.

This approach is taxonomically similar to the hostnames used in
Toradex SoMs, which is always the machine name (that as in the latest
paragraph also has the format of <family>-<soc>, such as verdin-am62.

With this approach, the differences between hostnames, device ids and
Yocto machines is unequivocal and clear.

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>