docs: rework build instructions section
Reorder the "Building Common Torizon OS" section to improve user
experience.

The machine-specific instructions are now presented first, as this is
the most common starting point for users. Generic instructions are moved
to a subsection.

The introductory text has also been shortened to be more direct and
concise, in line with a developer-focused audience.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>