initramfs-framework: refactor initramfs graphical kernel modules
To avoid creating various .conf files for each module configuration
necessary to have splash screen, we now have INITRAMFS_EXTRA_KMODS
variable.
This variable holds the name of the modules needed to be loaded by the
initramfs for splash screen to work. And the modules specified there are
also probed in that order, so this is very important.

And to reduce noise in the .bbappend recipe, this graphical modules
dependencies will be now tracked in the 'initramfs-graphics.inc' include
file.

Related-to: TOR-3894

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit 0930f87e0e5b4ad75ccbe98697c2c24647a22261)
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>