linux-firmware: add missing firmware for Realtek 8723
The firmwares for Realtek 8723 are already present in linux-firmware
repository. Only the Scarthgap recipe for linux-firmware that is not
installig it (on master it has already been added).

So we add this .bbappend to install the missing firmware binaries.

Related-to: TOR-3993

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit 348ed32ab75ca3a6e82cc77cc3daa2e327473aee)