ostree: add patch to save torizon_boot_args from previous deployment
This will be used as part of a new method to allow adding custom
kernel arguments in Torizon OS. This will save the value of
torizon_boot_args in uEnv.txt from the previous deployment,
allowing the custom kernel args to be tracked via OSTree.

This only works with 2 deployments the current and previous.

Related-to: TOR-4066

Signed-off-by: Jeremias Cordoba <jeremias.cordoba@toradex.com>