setup-environment-synaptics: prevent grep from finishing the script
On our Jenkins CI, our 'setup-environment' script are being called in
a bash set with '-e' option.

Since we grep 'conf/local.conf' for the accepted licenses, it will fail
and cause bash to exit. But this is expected behavior.

So, we add '|| true', to not cause early termination of the script, and
it also won't affect the flow of the first use scenario.

Related-to: TOR-4068

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>