aktualizr-default-sec: fix bash conflict when generating SDK
When generating the SDK, yocto uses 'target-sdk-provides-dummy' to
accept the recipes dependencies, but without adding tools that are
already present in the native sysroot.

But the issue here was that, the scripts were using the
"#!/usr/bin/bash" shebangs, which are not specified in
'target-sdk-provides-dummy', but on 'bash' recipe. And since
'target-sdk-provides-dummy' is set to conflict with 'bash' this issue
arise. Because when DNF is solving the dependencies,
'target-sdk-provides-dummy' won't satisfy, so the dummy mechanism won't
take place, instead it will find a dependency with 'bash' recipe, which
will cause a conflict, by design, with 'target-sdk-provides-dummy'.

The solution here was to change the shebang to "$!/bin/bash", since this
is provided by 'target-sdk-provides-dummy', and the correct dependencies
will be formed.

Related-to: TOR-3950

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit 2d4b34989126cee3f51da97fb67a2a10e3f4478d)
Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>